requirements.txt
streamlit
python-docx
beautifulsoup4
requests
openai>=1.0.0
python-dotenv
PyPDF2>=3.0.0